Inventory Count Report
This is a quick Inventory Count Report for your Fishbowl Inventory users.  This will create an on-screen table listing all inventory, and give them the ability to export to Excel or PDF. There is a similar report in Fishbowl but this is much quicker and does not require an additional login.  The uikit-main.css and uikit-main.js are from the UIKit framework.